# docs/public/api/swagger-2.0.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Track, Manage, and Organize your Things.",
        "title": "Homebox API",
        "contact": {
            "name": "Homebox Team",
            "url": "https://discord.homebox.software"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/v1/groups/statistics": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Statistics"
                ],
                "summary": "Get Group Statistics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/repo.GroupStatistics"
                        }
                    }
                }
            }
        },
...
        "/v1/groups/statistics/locations": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Statistics"
                ],
                "summary": "Get Location Statistics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/repo.TotalsByOrganizer"
                            }
                        }
                    }
                }
            }
        },
        "/v1/groups/statistics/purchase-price": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Statistics"
                ],
                "summary": "Get Purchase Price Statistics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "start date",
                        "name": "start",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "end date",
                        "name": "end",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/repo.ValueOverTime"
                        }
                    }
                }
            }
        },
        "/v1/groups/statistics/tags": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Statistics"
                ],
                "summary": "Get Tags Statistics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/repo.TotalsByOrganizer"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "repo.GroupStatistics": {
            "type": "object",
            "properties": {
                "totalItemPrice": {
                    "type": "number"
                },
                "totalItems": {
                    "type": "integer"
                },
                "totalLocations": {
                    "type": "integer"
                },
                "totalTags": {
                    "type": "integer"
                },
                "totalUsers": {
                    "type": "integer"
                },
                "totalWithWarranty": {
                    "type": "integer"
                }
            }
        },
...
        "repo.TotalsByOrganizer": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "total": {
                    "type": "number"
                }
            }
        },
...
        "repo.ValueOverTime": {
            "type": "object",
            "properties": {
                "end": {
                    "type": "string"
                },
                "entries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/repo.ValueOverTimeEntry"
                    }
                },
                "start": {
                    "type": "string"
                },
                "valueAtEnd": {
                    "type": "number"
                },
                "valueAtStart": {
                    "type": "number"
                }
            }
        },
        "repo.ValueOverTimeEntry": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "value": {
                    "type": "number"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "\"Type 'Bearer TOKEN' to correctly set the API Key\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}